+++ src/Sparkle/CSharp/Physics/Simulation.cs
@@ -8,5 +8,5 @@ public abstract class Simulation : Disposable {
     /// Performs a single step in the physics simulation based on the given time step.
     /// </summary>
     /// <param name="timeStep">The duration of the step in seconds.</param>
-    protected internal abstract void Step(float timeStep);
+    protected internal abstract void Step(double timeStep);
 }
